requirements.txt
Flask==3.1.0
transformers==4.46.3
torch==2.5.1
scipy==1.14.1
spacy==3.7.5
coqui-tts==0.24.3